Move the value diff to task level
